♻️ Refactor: 검색 입력 및 라우팅 로직 커스텀 훅으로 분리 (#5)
